+++ RELEASE-NOTES.md
@@ -31,6 +31,7 @@
 * Changing imports in function/procedure section in `snowflake.yml` will cause the definition update on replace
 * Adding `--pattern` flag to `stage list` command for filtering out results with regex.
 * Fixed snowpark build paths for builds with --project option (fixed empty zip issue).
+* Adding support for any source supported by `pip` in `snow snowpark`.
 
 # v2.1.2